Remove pim_catalog_product_identifiers table and calls
