+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -142,11 +142,12 @@ services:
         tags:
             - { name: 'kernel.event_subscriber' }
 
-    Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Db\CreateIdentifiersTableSubscriber:
-        arguments:
-            - '@database_connection'
-        tags:
-            - { name: 'kernel.event_subscriber' }
+#   TODO CPM-1102: Re-enable when table can be created again
+#    Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Db\CreateIdentifiersTableSubscriber:
+#        arguments:
+#            - '@database_connection'
+#        tags:
+#            - { name: 'kernel.event_subscriber' }
 
     pim_catalog.event_subscriber.attribute_option_set_sort_order:
         class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\AttributeOption\SetAttributeOptionSortOrderSubscriber'
